backend/requirements.txt
modal>=0.64.0
fastapi>=0.110.0
uvicorn>=0.29.0
google-generativeai>=0.5.0
httpx>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.9
